move emoji board component hooks in hooks dir
