+++ src/app/components/emoji-board/index.ts
@@ -1,2 +1,4 @@
 export * from './emoji';
+export * from './custom-emoji';
+export * from './recent-emoji';
 export * from './EmojiBoard';
