+++ src/app/hooks/useRecentEmoji.ts
@@ -1,8 +1,7 @@
 import { useEffect, useState } from 'react';
 import { ClientEvent, MatrixClient, MatrixEvent } from 'matrix-js-sdk';
-import { IEmoji } from './emoji';
-import { getRecentEmojis } from './recent-emoji';
-import { AccountDataEvent } from '../../../types/matrix/accountData';
+import { IEmoji, getRecentEmojis } from '../components/emoji-board';
+import { AccountDataEvent } from '../../types/matrix/accountData';
 
 export const useRecentEmoji = (mx: MatrixClient, limit?: number): IEmoji[] => {
   const [recentEmoji, setRecentEmoji] = useState(() => getRecentEmojis(mx, limit));
